Convert to MPP module, looks like writing to vertex buffer from MPPBuffer is broken on JVM?
