+++ orx-mesh-generators/src/jvmDemo/kotlin/DemoBox.kt
