+++ orx-mesh-generators/src/jvmDemo/kotlin/DemoComplex01.kt
@@ -3,14 +3,11 @@ import org.openrndr.application
 import org.openrndr.color.ColorRGBa
 import org.openrndr.draw.DrawPrimitive
 import org.openrndr.draw.shadeStyle
-import org.openrndr.extensions.SingleScreenshot
 import org.openrndr.extra.camera.Orbital
 import org.openrndr.extra.meshgenerators.box
-import org.openrndr.extra.meshgenerators.group
 import org.openrndr.extra.meshgenerators.meshGenerator
 import org.openrndr.extra.meshgenerators.sphere
 import org.openrndr.math.Vector3
-import org.openrndr.math.transforms.transform
 
 fun main() {
     application {
