+++ orx-mesh-generators/src/jvmDemo/kotlin/DemoComplex03.kt
