+++ orx-mesh-generators/src/jvmDemo/kotlin/DemoComplex04.kt
