+++ orx-mesh-generators/src/jvmDemo/kotlin/DemoComplex05.kt
